worker/requirements.txt
pydantic-settings
redis
pymongo
motor
docling
weaviate-client
python-dotenv
bullmq
colorama
motor
protobuf<=5.28.3
grpcio==1.67.1
loguru
beanie
aiohttp
openai
pdf2image
poppler-utils
pypdf
Pillow
minio
py7zr
rarfile
patool
